requirements.txt
# Google Calendar API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dateutil>=2.8.2

# Display & Image rendering
Pillow>=10.0.0

# Database
sqlite3-python>=1.0.0

# Waveshare e-paper driver
# (For RPi only - skipped on development machine)
# spidev>=3.6
# RPi.GPIO>=0.7.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
